REFACTOR storage service to improve testability and maintanability
