+++ .github/workflows/build-and-publish.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: [3.6, 3.7, 3.8, 3.9]
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python
@@ -28,8 +28,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install setuptools wheel twine
         pip install -r requirements.txt
-        mkdir test-reports
-        nosetests -s --with-xunit --xunit-file=test-reports/test-report.xml
+        pytest --junitxml="test-reports/test-report.xml"
   deploy:
     needs: build
     runs-on: ubuntu-latest
@@ -45,7 +44,7 @@ jobs:
         TWINE_PASSWORD: ${{ secrets.PYPI_PASSWORD }}
       run: |
         pip install -r requirements.txt
-        pip install semversioner==0.8.1
+        pip install semversioner==0.10.0
 
         previous_version=$(semversioner current-version)
         semversioner release
