+++ .github/workflows/build-and-test.yml
@@ -9,7 +9,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: [3.6, 3.7, 3.8]
+        python-version: [3.6, 3.7, 3.8, 3.9]
     steps:
     - uses: actions/checkout@v2
     - name: Set up Python
@@ -27,5 +27,4 @@ jobs:
         python -m pip install --upgrade pip
         pip install setuptools wheel twine flake8
         pip install -r requirements.txt
-        mkdir test-reports
-        nosetests -s --with-xunit --xunit-file=test-reports/test-report.xml 
+        pytest --junitxml="test-reports/test-report.xml"
