+++ requirements.txt
@@ -1,4 +1,4 @@
-nose==1.3.7
+pytest==6.2.1
 click==7.1.2
 twine==3.2.0
 wheel==0.34.2
